# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server TC-DAPP server.",
        "title": "tcDAPP APIs",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/dapp-moderator/v1",
    "paths": {
...
        "/token-explorer/token/{address}": {
            "get": {
                "description": "Update token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "token-explorer"
                ],
                "summary": "Update token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "contractAddress",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.JsonResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Get token detail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "token-explorer"
                ],
                "summary": "Get token detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "contractAddress",
                        "name": "address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "contractAddress",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateTokenReq"
                        }
                    },
                    {
                        "type": "file",
                        "description": "thumbnail file",
                        "name": "thumbnail_file",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.JsonResponse"
                        }
                    }
                }
            }
        },
        "/token-explorer/tokens": {
            "get": {
                "description": "Get tokens",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "token-explorer"
                ],
                "summary": "Get tokens",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "page",
                        "name": "key",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.JsonResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "request.UpdateTokenReq": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "social": {
                    "type": "object",
                    "properties": {
                        "discord": {
                            "type": "string"
                        },
                        "instagram": {
                            "type": "string"
                        },
                        "medium": {
                            "type": "string"
                        },
                        "telegram": {
                            "type": "string"
                        },
                        "twitter": {
                            "type": "string"
                        },
                        "website": {
                            "type": "string"
                        }
                    }
                },
                "symbol": {
                    "type": "string"
                },
                "thumbnail": {
                    "type": "string"
                }
            }
        },
        "response.JsonResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/response.RespondErr"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "response.RespondErr": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}